+++ src/xnnpack/simd/u8-sse2.h
@@ -72,7 +72,7 @@ static XNN_INLINE void xnn_storeu_u8(uint8_t* ptr, xnn_simd_u8_t v) {
   _mm_storeu_si128((__m128i*)ptr, v);
 }
 
-static XNN_INLINE void xnn_store_u8(float* ptr, xnn_simd_u8_t v) {
+static XNN_INLINE void xnn_store_u8(uint8_t* ptr, xnn_simd_u8_t v) {
   _mm_store_si128((__m128i*)ptr, v);
 }
